[system] Rename types
This caused PyPI to fail
